move TypedChannel to crate::channel
